[refact]: credit ui 변경
